ai-service/requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0
opencv-python>=4.7.0
albumentations>=1.3.0
tensorboard>=2.10.0
tqdm>=4.64.0
efficientnet-pytorch>=0.7.1
timm>=0.6.12

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
